+++ api/src/errors/error.ts
@@ -1,5 +1,5 @@
-import { LLMProvider } from 'api/types.ts';
-import { ConversationId } from 'shared/types.ts';
+import type { LLMProvider } from 'api/types.ts';
+import type { ConversationId } from 'shared/types.ts';
 export type { ErrorObject as AjvErrorObject } from 'ajv';
 import { Status } from '@oak/oak';