+++ api/src/llms/llmMessage.ts
@@ -1,27 +1,36 @@
-import {
+import { ulid } from '@std/ulid';
+import type {
 	LLMMessageStop,
 	LLMProviderMessageResponseRole,
 	LLMProviderMessageResponseType,
 	LLMTokenUsage,
 } from 'api/types.ts';
-import { LLMToolInputSchema } from 'api/llms/llmTool.ts';
+import type { LLMToolInputSchema } from 'api/llms/llmTool.ts';
 
 export interface LLMMessageContentPartTextBlock {
+	messageId?: string;
 	type: 'text';
 	text: string;
 }
 
 export interface LLMMessageContentPartImageBlock {
+	messageId?: string;
 	type: 'image';
 	source: LLMMessageContentPartImageBlockSource;
 }
+export type LLMMessageContentPartImageBlockSourceMediaType =
+	| 'image/jpeg'
+	| 'image/png'
+	| 'image/gif'
+	| 'image/webp';
 export interface LLMMessageContentPartImageBlockSource {
 	data: string;
-	media_type: 'image/jpeg' | 'image/png' | 'image/gif' | 'image/webp';
+	media_type: LLMMessageContentPartImageBlockSourceMediaType;
 	type: 'base64';
 }
 
 export interface LLMMessageContentPartToolUseBlock {
+	messageId?: string;
 	type: 'tool_use' | 'tool_calls'; // tool_use is anthropic - tool_calls is openai
 	id: string;
 	input: object;
@@ -29,14 +38,23 @@ export interface LLMMessageContentPartToolUseBlock {
 }
 
 export interface LLMMessageContentPartToolResultBlock {
+	messageId?: string;
 	type: 'tool_result' | 'tool'; // tool_result is anthropic - tool is openai
 	tool_use_id?: string; // anthropic
 	tool_call_id?: string; // openai
-	content?: Array<LLMMessageContentPartTextBlock | LLMMessageContentPartImageBlock>;
+	content?: Array<
+		LLMMessageContentPartTextBlock | LLMMessageContentPartImageBlock
+	>;
 	is_error?: boolean;
 }
 
-export type LLMMessageContentPartType = 'text' | 'image' | 'tool_use' | 'tool_calls' | 'tool_result' | 'tool';
+export type LLMMessageContentPartType =
+	| 'text'
+	| 'image'
+	| 'tool_use' // anthropic
+	| 'tool_result' // anthropic
+	| 'tool_calls' // openai
+	| 'tool'; // openai
 
 export type LLMMessageContentPart =
 	| LLMMessageContentPartTextBlock
@@ -79,7 +97,14 @@ class LLMMessage {
 		public providerResponse?: LLMMessageProviderResponse,
 		public id?: string,
 	) {
+		this.setId();
 		this.setTimestamp();
+	}
+
+	public setId(): void {
+		if (!this.id) {
+			this.id = ulid();
+		}
 	}
 
 	public setTimestamp(): void {